+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Add `CircleCreate` command to the example project.
 - Add `MessageBox` and `ShowBalloon` buttons to the example ribbon panel.
 - Update resource image to `Cube-Grey-Light.tiff` and `Cube-Grey-Dark.tiff`.
+- Add `AutoCADBusyService` to update the button state in the example project.
 ### Fixes
 - Fix `CheckURLValid` to support `http` and `https`.
 - Fix `SetImage` to update `Image` property.